requirements.txt
anyio==4.6.2.post1
click==8.1.7
fastapi==0.95.2
graphql-core==3.2.5
h11==0.14.0
idna==3.10
pydantic==1.10.18
python-dateutil==2.9.0.post0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
strawberry-graphql==0.217.1
typing_extensions==4.12.2
uvicorn==0.32.0
